backend/requirements.txt
flask
flask-cors
requests
python-dotenv
yfinance>=0.2.32
